Greatly simplify and de-duplicate code
